Update patch set 2
Patch Set 2: Code-Review+1

I've seen more than one report that this change helped with reboots on bacon.  I am not sure why but the selinux denials right before the kernel panics are a good indicator.  Also, I had some people test a build with this policy in it and their reboots have dropped or not occured since then.  I think this is onto something.

Patch-set: 2