Update patch set 1
Patch Set 1:

Can you globally disable jack?  I think the issue is that anything depending on the jack file won't cause a rebuild of the library prebuilt module, so it needs to be generated at the library, and in both formats because AFAIK disabling/enabling jack is on a per module basis.

Patch-set: 1