Update patch set 1
Patch Set 1:

> that seems like a different issue.

Yeah, I too think it's a different issue, but I thought it was worth to point it out. I used DriveDroid with cm-12.1, so there must have been some change that broke it. I +1d this because I think it's needed in general to allow the use of the feature, which is not limited to DriveDroid.

> Did the DriveDroid image pop up on your PC when you've plugged in your phone?

Only during the setup.

> Regarding the current problem, could you send me a support email from within DriveDroid?

Done. We can keep discussing somewhere else about the current issue.

Patch-set: 1