Update patch set 1
Patch Set 1:

I can certainly merge this local to the device, but it seems more logical to put it in the place we set all the other interactive governor permissions.

Note this node is not present on all devices, depends on what set of interactive governor patches are on the device.

Patch-set: 1