Update patch set 1
Patch Set 1:

User report is both right and wrong. 

Datos has both a mnc 34 and 341. Changing the 34 outright is not the proper solution. Instead, make sure we conform to what aosp provides as the settings for these. 

Further, the user's own assessment that when changing the mnc to 341 the apn 'disappears' means that they do in fact have a mnc 34 designated sim.

Patch-set: 1