Update patch set 1
Patch Set 1:

@Steve Kondik:
I don't know whether I'm understanding your question. DriveDroid needs to access su in order to execute shell commands under root. By doing that it will instruct the kernel to use a certain file (usually a .ISO file) for its USB Mass Storage gadget. By doing so, the PC will see a USB mass storage device with the contents of the .ISO file. This allows a PC to boot from a Android phone.

Among others, the following commands are executed by DriveDroid, as Gabriele mentioned. These fail without the su.te rule in this patch:

  echo 0 > /sys/class/android_usb/android0/enable
  echo MYISO > /sys/devices/platform/msm_hsusb/gadget/lun0/file
  echo mass_storage > /sys/class/android_usb/android0/functions
  echo 1 > /sys/class/android_usb/android0/enable

where MYISO is an file that resides on the internal/external storage.

These commands execute successfully, but after that the kernel will try to read from MYISO. This fails with the current policies on CM13.

I hope that clarifies the need for this patch.

Just a side-note: the app doesn't seem to suffer from this problem on other roms (including stock 6.x roms). It's presumably because SuperSU 2.11+ has workarounds for this. SuperSU actively patches SElinux policies.

Patch-set: 1