Update patch set 14
Patch Set 14:

Whole bunch of fixes to simplify things. This changes a few things (for the better) in the top level makefile as well. The "setup_vendor" function should work for handling the device-common case and there is no more requirement that global variables are inherited. I also fixed the lazy-bastard-line-endings thing :)

Patch-set: 14