Update patch set 24
Patch Set 24:

Is it safe to use Bash 4 syntax? I've heard this might be a problem for Macs. If not, I feel like this is going to get incomprehensible soon because you can't pass an array by reference on Bash 3 and this whole thing could be done in 1/3 LOC with Python.

Patch-set: 24