Update patch set 28
Patch Set 28:

The PAIN!

New patchset reworks a whole bunch of stuff so we can deal with an additional argument for each line in the file. Currently this is used for passing the name of the certificate which should be used to sign the APK. Additionally, this should work when filenames have spaces, if that's a thing.

Patch-set: 28