Update patch set 29
Patch Set 29:

I feel bad for anyone trying to debug this. Passing the name of a local array variable as a function argument and reconstituting it using indirection is absolutely horrible. What the fuck were the authors of Bash thinking?

I'll redo this in Python at some point, but we're super close now so let's finish it.

Patch-set: 29