Update patch set 2
Restored

I'll un-abandon, but I don't think this is the route we really want to take to fix the issue. VZWADMIN becomes the default when this APN is gone, which admittedly is probably ok since it doesn't specify mms as one of its types. I've so far been just too damn lazy to see what changed in APN selection between cm12.1 and cm13.

Patch-set: 2
Status: new