Update patch set 2
Patch Set 2: Code-Review-1

Actually, the bug might be with PackageManager. It should install the new app to /mnt/expand/guid/, but it installs it to /mnt/asec/, which is a tmpfs and thus it disappears during reboot. This might be the correct course of action (I'm not sure, just speculating), but at some point it fails and thus fails to move the file from /mnt/asec/ back to /mnt/expand.

Here are some interesting bits from the log:

07-19 21:40:10.906: I/PackageManager(673): Package com.myapp codePath changed from /mnt/expand/ef85ad67-e334-4679-a197-c890ffaa350d/app/com.myapp-1 to /mnt/asec/com.myapp-2; Retaining data and using new

07-19 21:40:11.403: W/PackageManager(673): Couldn't remove dex file for package:  at location /mnt/expand/ef85ad67-e334-4679-a197-c890ffaa350d/app/com.myapp-1/base.apk, retcode=-1

07-19 21:40:11.581: D/PkgBroadcastIntentOp(1417): Received broadcast action=android.intent.action.PACKAGE_REMOVED and uri=com.myapp

07-19 21:40:11.644: E/Drive.UninstallOperation(1417): Package still installed com.myapp

07-19 21:40:12.001: D/PkgBroadcastIntentOp(1417): Received broadcast action=android.intent.action.PACKAGE_ADDED and uri=com.myapp

07-19 21:40:13.352: D/PkgBroadcastIntentOp(1417): Received broadcast action=android.intent.action.PACKAGE_REPLACED and uri=com.myapp

Patch-set: 2
Label: Code-Review=-1