Update patch set 6
Patch Set 6:

I'd recommend putting all the "hacks" like this into a separate file and sourcing it from the main functions. The only reason I suggest this is because it's already a little nasty to look at, and breakage will potentially break lots of devices.

Patch seems fine, but I can think of at least one case where this will cause damage- some versions of the QDCM tool produce the same invalid XML. The problem is that if you correct it like this, it will crash the library! I can't even make this stuff up. They fixed it in newer versions of the code (now malformed XML crashes it), but it's hard to say what's out there right now.

Patch-set: 6