Update patch set 27
Patch Set 27:

> Works great with smali 2.1.2
 > Tested with adb and directory sources on angler.
 > 
 > Should we add a version check (not sure what minimum compatible
 > version is) or auto-checkout+gradle step?

If you ask me, I would prefer shipping baksmali and smali .jars in vendor/cm. That would avoid the user setting up additional environment flags, simplify the code (we could get rid of the checks) and yes, it would ensure everyone was using the same version of the tools.

Waiting for Steve to step in.

Patch-set: 27