Update patch set 30
Patch Set 30:

Maybe I missed it, but I didn't see a response to Bruno Martins' inquiry "I don't see the point. Why wouldn't the user want to deodex the apk|jar?"

In my experience, the blobs put in vendor are either decoded->recompiled->signed already, or at least decoded->recompiled and then the build system signs them. The blobs extracted by extract_utils should match the blobs already in vendor. Is this not the case on angler (and/or others)?

Patch-set: 30