Update patch set 2
Patch Set 2:

I don't disagree with "The TARGET_VENDOR_SHOW_MAINTENANCE_VERSION should act as it says".  Unfortunately, it's more important to meet the requirement we have.  I'd prefer to maintain the logic within the build system to allow dropping of the maintenance version if == 0.

That being said, if you put up a patch that meets the original requirements and is more logical, I'm happy to review it.

Patch-set: 2