Update patch set 14
Patch Set 14:

There is no way around mounting data to get the time when qcom does that stupid thing. Also the timezone is there so it's required to get the correct one.
As also discussed on IRC I don't agree about staying away from a feature because not everyone might be able to use it. The scenario Non-encrypted->Encrypted->"Where is the time now" is one I don't think will occur often enough to stay away from this

Patch-set: 14