Update patch set 3
Patch Set 3:

All tools represent a total of 4MB and required so should be included.
They could as well be part of common.mk rather than duplicated.
Unless there's something we don't need in mini, though 4MB...

The apps added to mini could be moved to common too while we're at it.

Also, 3-5MB but do we really need the cm audio to be included in mini too,
knowing it's duplicated there I imagine it was for a reason.

I had in mind to add a set like common_normal.mk that would include
the current common_mini stuff, and let the mini be really minimal
for only what is needed to use for lightweight devices.

Patch-set: 3