Update patch set 2
Patch Set 2:

Fwiw this commit doesn't fix Google music being able to store on an external sdcard when sdcardfs is enabled, which was my initial reason for starting to fix this, but it does fix general access to both internal and external. Google music crashes on N with sdcardfs external storage too, so probably their app just sucks.

Patch-set: 2