Update patch set 2
Patch Set 2: Code-Review-1

I don't know why the denial as it happens before the service's constructor is called.  It's a non-issue I have not bothered to look into it.  The denial does not cause any adverse effects so not a big priority in figuring out why the denial.

Patch-set: 2
Label: Code-Review=-1