Update patch set 1
Patch Set 1:

Where does the file end up? Also, I'm not if sure this is a good idea. If I pull in a bunch of patches and the build fails due to new config items, I probably should go and take a look to be sure I didn't forget anything.

On a similar note, we need to figure out why Ninja rebuilds everything with a kernel dependency every time.

Patch-set: 1