Update patch set 2
Patch Set 2:

sam mentioned that maybe writing a helper script to generate these so we're not modifying these at compile might be a better solution.
If anyone has taken a look at this, you probably noticed we're using 1.6 universally for a width-to-height ratio. We can pass the height as a flag to the script, but then would that break the ratio of the images we have is my question. thoughts?

Patch-set: 2