Update patch set 1
Patch Set 1:

The proper fix depends on what we want the behaviour to be. The code seems to demand the following:

There must be at least one WebView package that is available by default.
There must be at least one WebView package that is available by default and not a fallback.
Settings a fallback is optional but that package will be disabled if another non-fallback package is available.

Do we want our source-built WebView to be disabled when another one in the list is available?

Patch-set: 1