Update patch set 1
Patch Set 1:

unsure if you want it like this, works for me though.

This would also allow to configure a different path per arch, though we might want to configure that differently (maybe build system instead of envsetup?)

Anyways, that gets building started for me on macOS, if the SDCLANG dirs are present.

Patch-set: 1