Update patch set 2
Patch Set 2:

It looks like the hardware/qcom/media pathmap got screwed up in 14.1. The pathmap needs to point to the hardware/qcom/media/msm8996|msm8974 depending on your platform. Right now it points at hardware/qcom/media, which defeats the point of the pathmap, because any user of the AOSP pathmap is required to specify $(call project-path-for,qcom-media)/msm8974 when they retrieve the project path.

Patch-set: 2