Update patch set 6
Patch Set 6:

> Seems OK, but shouldn't the presence of BOARD_CACHEIMAGE_PARTITION_SIZE
 > mean "hey guyz I have a cache partition and I want to use it"..?

Yea, I guess either would mean that though. I opt'd for this since in system/core/rootdir/Android.mk its using this flag to determine whether or not it should mkdir /cache or symlink it to /data/cache in the ramdisk, whereas the size param is only used in build/ to determine how big to make the cache.img.

Patch-set: 6