Update patch set 1
Patch Set 1:

It will break it for those that haven't enabled the TARGET_USES_MEDIA_EXTENSIONS flag. Either due to their respective media HAL doesn't have support for it, or that they can't use it because of their camera HAL. Enabling the flag will trigger the condition in the patch I linked to above, forcing the use of native handle buffers instead of gralloc.

Personally, since 8960 media HAL now has support for it, I can enable it and regular recording works fine, despite that my camera definitely don't have support for it. Magic? Dunno. Only thing not working is secure encoding, which is why I can't switch to it fully yet.

Patch-set: 1