Update patch set 2
Patch Set 2:

I see the same as @Adrian DC on two differnt devices running CM13.

For some reason batterymanager is running in the untrusted_app context, so the platform_app rule added to platform_app.tm doesn't have any effect.

So, why is the batterymanager service running in the untrusted_app
context instead of the platform_app context?  That seems to be the
actual problem here.

Patch-set: 2