Update patch set 3
Patch Set 3: Code-Review-2

Purely for debugging. Its a lot easier to have a dedicated place for audit to log denials. For instance if I want to check what's been denied in the last couple days. This couples with adding luca's auditd repo to your manifest. But we definitely shouldn't ship this.

https://github.com/luca020400/android_external_auditd

Patch-set: 3