Update patch set 1
Patch Set 1:

to give all of these commits some context, the purpose of them is to be able to package all executables and shared libraries and when doing so.

The reason for that is that this way not only can you link against any library, the dependencies can also be checked at compile-time and you don't need to wait for the linker errors to happen during runtime.

I recommend making packages for all props for all devices by adding the '-' prefix to the prop lists in front of every executable and library.

Patch-set: 1
Reviewer: Michael Zimmermann <5842@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0