Update patch set 2
Patch Set 2: Code-Review-1

Unlike the situation with 8974 and 8226, 8996 and 8953/8937 use different subdirs in most of the HALs. (8937/8953 use the 8916 subdir, and 8996 uses 8974) This would make the situation kind of screwy still, where we end up in the situation where we have to manually pick changes from within one branch to the same branch.

Patch-set: 2