Update patch set 2
Patch Set 2:

Maybe I can explain it better in words. For the LineageOS copyright if/elses:

The first if/elif pair is used for the cases when there is not a year range. The only time this will happen is this year (2017) and when the initial year matches the current year.

The last two elif/else handle the cases when there is a year range. If the initial year is <= 2017, then LineageOS copyright starts with 2017, otherwise if the initial year is greater than 2017, that year starts the range.

Patch-set: 2