envsetup: Add mkal to build modules using limited CPU resources
mkal works similarily to mka, but limits the CPU usage to the given
percentage value. This allows to run long lasting builds as
background tasks without taking all the resources available.

The commands can only be used with Linux and requires
CONFIG_CGROUP_SCHED and CONFIG_CFS_BANDWIDTH to be enabled.
It also requires 'sudo' and 'cgcreate' to create the cgroup and
'cgexec' to start 'make' in the cgroup.

Change-Id: I285b929857a058c733baa7a7313fe4debed8af81