Update patch set 8
Patch Set 8:

So no systemless boot in the boot image? Honestly, I'm still in-favor of just merging this and making people flash the su add-on as a one time deal. Like I said it would also be amazing to get support for addon.d to actually flash scripts you put in the folder or just like I talked about with another dev adding the ability in the Updater to have extra zips to flash after the OTA like some of the more popular apps. I know of multiple people that would love this feature and it would be very useful even if it's hidden by default.

Patch-set: 8