Update patch set 1
Patch Set 1:

> A better message would definitely clear up the intention here... 
 > And, fwiw, I would be happy to remove it given that there are
 > numerous options available that are better supported and
 > maintained.

For the most part, the "Downloads" launcher is sufficient. Where it falls short is that it doesn't show the root of internal storage, just the Images, Videos, and Audio directories. To reach the root of internal storage using only built-in tools, one must go to Settings > Storage (Internal shared storage) > Explore. My preference would be that if/when CMFileManager is removed, it is replaced with a storage explorer launcher (as described above). I'm impartial as to whether the Downloads launcher remains if this is done.

I tried my hand at writing my first app that would simply launch this action, but I haven't gotten it to work yet. Someone who deals with apps could probably whip this up pretty quick (and could even hijack CMFileManager's folder icon if feeling lazy).

Patch-set: 1
Label: Verified=0