Update patch set 3
Patch Set 3:

More changes == more things can do wrong

and this will be precisely at a time where it HAS to work.

If this is for key migration, I would say just do it in the existing script with as few lines as possible.  We can always replace the shell script with your cpp version later once the critical period migration period is over.

Patch-set: 3