Update patch set 3
Patch Set 3: Code-Review-1

No, this is not for the migration, it's just something I wanted to do. As I said, I had quite some code ready that I simply reused.

I just find the C++ version simpler and more robust, mostly because it only deals with the navigation of the data and delegates the parsing of the xml to a well tested library.

I agree with you, this can wait and I didn't upload it to get it merged right away.

-1 for now, there are some changes I want to do that I'll do when I can.

Patch-set: 3
Reviewer: Gabriele M <13897@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=-1