Update patch set 9
Patch Set 9:

There are plenty of alternatives until the official method is released. But there are no official builds so I'm pretty sure they're not going to release this change when technically you shouldn't even be on a build yet. I've been using the phh su for months without issue. There are plenty of options that work just fine until official release which will most likely have the add-on released along side it.

Patch-set: 9