Update patch set 6
Patch Set 6: -Code-Review

I'm thinking of doing things a bit differently: I'd include 'su' in every build (in /install/extra maybe) and replace 51-addonsu.sh with a 'flag file'. If we find the 'flag file' while flashing a new weekly, we copy the latest version of 'su' in /system/xbin/su. I think this allows to easily update the binary in case we need to.

Patch-set: 6
Label: Code-Review=0