Update patch set 3
Patch Set 3:

That's why I'm asking because I am the proofreader for Romanian translation (since 11 January) and there was no other proofreader since Lineage was available for translation. I began with translation from 58% and there was nobody to translate this language (now it's fully translated and approved). My question is: am I allowed to have my name in the list even if the dump is from 30 Dec.?

Patch-set: 3
Label: Verified=0