Update patch set 5
Patch Set 5:

Editing stuff on gerrit is just the worst....

Just moved the file.te and file_contexts entries to the top so hopefully less conflicts will arise with this commit and other things that we may be testing. Since this commit isn't intended to be merged, I really don't care about the order.

Patch-set: 5