Update patch set 3
Patch Set 3: Code-Review-2

We discussed this already, and came to the conclusion that it should be fine to remove it after improving DocumentsUI to be obvious from the launcher that it's not downloads only. I don't recall anymore what communications channel that was on :-/ but Joey should remember the details. As far as I can tell, this hasn't happened yet.

Patch-set: 3