Update patch set 14
Patch Set 14:

> Do we plan on committing this patch any time soon?

I've just noticed that backuptool is not reliable to save/restore SELinux contexts. Depending on the version of 'cp' used, you have to use a different syntax to preserve the SELinux context of a file. The context may be applied automatically (not sure of this), but that's not to be taken for granted and we still depend on the recovery's file_contexts for this.

Patch-set: 14
Reviewer: Gabriele M <13897@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0