backuptool: Move the files rather than copying them
When files are moved all the attributes are preserved. 'cp'
has different command line arguments to preserve attributes, but
the syntax depend on the specific on the binary used, so use 'mv'
which requires no arguments. Preserving a copy of the original
is unnecessary unless there's some failure right before /system
is updated.

Change-Id: If2c245863df5675c18dbf43b6bcedeb33383fc38