Update patch set 1
Patch Set 1:

I thought part of the joke was that it was merged a day late, but each day it remains a part of the codebase makes it less funny. I genuinely freaked out when I first saw the off-center squid/transistor and br0zip app full of spelling mistakes, but it got annoying fast. It was easy enough to manually revert in my synced repo.

Patch-set: 1