Update patch set 1
Patch Set 1:

Enough with the +1ing. This is not Facebook. The number of plus ones is NOT cumulative and does NOT mean it will be more likely to be merged. Reserve Code Review labels for technical reasoning ONLY.

This (alongside the bootanimation revert) will be merged after the LineageOS 14.1 builds have finished this week (Friday). 

If you can't deal with a simple harmless joke app, don't flash this week's builds.

Patch-set: 1