Update patch set 5
Patch Set 5:

> How do you want to test if the file exists on the device?

We are changing the way we get the files from the file system. The way we extract the files from the device is left unchanged, we always need the /system prefix in that case.

Currently if you want to extract the files from the filesystem, you must create a system/ directory and pass the path to its parent to extract-utils.


Just a minor correction on the snippet I posted, the first line of my snippet should be:

    elif [ -f "$SRC/$FILE" ] && [ "$SRC" != "adb" ]; then

Patch-set: 5
Reviewer: Gabriele M <13897@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0