build: Refactor remote prebuilt modules handling
Simplify the code and add the possibility to run builds without
an active internet connection.

To run an offline build, first pre-fetch all the remote prebuilt
modules running the following command:

    OFFLINE_BUILD=false mka fetchprebuilts

Then set the env variable OFFLINE_BUILD and start building:

    export OFFLINE_BUILD=true

A warning will be shown every time an offline build is started
to remind the user that prebuilt modules should be fetched.

Change-Id: I721cc5efbd3ea54d1d6c781f25adec0797e6b6ac