Update patch set 2
Patch Set 2:

> Xiaomi uses a bit different configs in their stock rom ~
 > https://hastebin.com/itugivetaw.scala
 > 
 > Wouldn't it be better to use that instead?

CTWAP is the old days wap APN
CTNET is for 3G
CTLTE is for 4G

This change is for CTLTE only, so it is OK to add CTWAP and CTNET.
The tiny difference about CTLTE is that: Some manufactures set CHAP and "internet", some don't. I have discussed this with a China Telecom staff. We confirmed that only apn="ctlte" is required, others are optional. I have been using this APN for a few days, everything works well.
You may open another change to add CTWAP and CTNET, but this CTLTE apn is correct.

Patch-set: 2