Add LTE support for China Telecom Macau.
Notice: I am very likely the only one here having a Macau SIM card, so I am the only one who can confirm this change.
Please do not paste APN settings copied from other vendors or devices. 
The exactly APN used in MIUI is not MMC455/MNC07 but MMC455/MNC02, maybe Xiaomi has done extra calculations to do so.

It is a waste of time to argue which APN is “BETTER”. This one works for China Telecom Macau LTE. I can confirm this. If you live out of Macau, you can't.

CTWAP and CTNET are not discussed in this change.
CTNET is for 3G. There are still someone using 3G. You can add one if you are interested in this, or I will add one if I have a confirmed APN settings from China Telecom Macau staffs.
CTWAP is outdated. If you use CTWAP in China, most apps will tell you "CTWAP is deprecated, use CTNET instead.".

China is moving from 3G to 4G/LTE, and will soon be moving from 4G to 5G. If you live out of China, please simply +1 and merge it. If you live in China and having problem with this APN, we can discuss this with beer and snacks.

For now, there is NO APN settings for MMC455/MNC07, so this APN won't harm anyone else living out of Macau.

Merge this change, that's it.

Change-Id: I8600dbbb8d3ea567f04a516d7a296dbaaebc5216