Update patch set 6
Patch Set 6: Code-Review-1

Move this all the way to the top of the XML, that's how it's supposed to be sorted. Also, I'm not sure that this won't cause issues, we'll have multiple "keys" with the same value in the map, which means they might overwrite each other :/ (if I'm understanding the code correctly). This is clever, but might be problematic, if that's the case. I'll do some testing later to see if that's the case

Patch-set: 6
Label: Code-Review=-1