Update patch set 7
Patch Set 7:

Regarding the incorrect signing of ctsshim (and whatever other packges need to be testkeys), should we keep a list of packages that need to be test-keys always? we could generate a list dynamically by cat packages | grep test-keys (I would think, at least. it's not elegant but w/e) and sed those lines back to test-keys if going to official

Patch-set: 7