Update patch set 3
Patch Set 3:

To go deeper. The APN list is serialized into the db in order of mcc/mnc pair. 

Type 'ia' is for initial activation, as is rightfully the top one in this scenario. 

Then in sequence, 'default,supl' is read for connectivity (including LTE), and finally, the 'mms' one is standalone for mms interactions. 

The ordering and data for each of these apn sets aligns with upstream AOSP - https://android.googlesource.com/device/google/marlin/+/android-7.1.1_r46/apns-full-conf.xml#9396

Finally, type 'internet+mms' is not a valid combination for type (to my knowledge: notice its lack of presence anywhere else in this file). 

For valid types, see 

https://github.com/LineageOS/android_frameworks_opt_telephony/blob/cm-14.1/src/java/com/android/internal/telephony/dataconnection/DcTracker.java#L1149

Patch-set: 3