Update patch set 1
Patch Set 1:

I wanted to do a similar change, so I agree with you. The extracted blobs are placed in vendor/ and stay there, there's no need to keep the content of system around in my opinion. I would actually store the files in the temp directory created by the script since it's automatically deleted at the end of the extraction.

Patch-set: 1
Label: Code-Review=0