kernel.mk: fail if `make modules' fails
Right now, the kernel build only fails if the kernel itself fails to
build.  The script ignores error while building modules.

This leads to a weird result:  Assuming a module patch introduces a
build error.  A subsequent bacon build will simply ignore the problem
and continue.  In the end, an installer zip file will be created, but
the affected module will be missing and nobody notices it until it
got installed.  Worse:  If the bacon build wasn't a clean build, the
broken module will still be installed, just the old version, without
anybody noticing.

This patch changes the `make modules' call so that it propagates a
failure to the calling make, thus making sure a failing module build
will be noticed immediately.

Change-Id: I3cc31f6948d5b03ad4371ad5448d34d4fdd67e0e
Signed-off-by: Corinna Vinschen <xda@vinschen.de>