cm: build: kernel: Rework kernel module logic
The existing kernel module logic does not detect failed kernel module
builds.  This is because the module build logic is a long shell chain
that invokes make macros which are not intended to be used in this way.
The essense of the issue is that we get a shell chain that looks like:

	cmd && cmd && var=value; cmd && var=value; cmd;

The shell breaks this into three separate chains.  The first is the one
that builds the modules.  The shell proceeds even if it fails.

The issue is made worse by the existing build rule conflating the built
kernel modules and the installed kernel modules.

Fix this by reducing the built kernel module command to a single shell
command and creating a separate target INSTALLED_KERNEL_MODULES.  Any
kernel module build failure will show up properly.  Further, the
commands for INSTALLED_KERNEL_MODULES do not invoke any macros and
avoid the issue described above.

Change-Id: I3d13056e217e1e937c8c3a345032682ffc394bca