Update patch set 2
Patch Set 2: Code-Review+1

I entered it exactly as the CVMovel agent entered it into my phone. I watched them configure a number of phones like this, so this must have been the 'least problematic'/most compatible variant they were taught to work with. I understand the function of this tag, and my APN worked with it set. No longer in a position to test this, since Cape Verdian SIM does not work in Europe to test further. Worst case: if it is reported to cause a problem, carrier_enabled can be removed.

Patch-set: 2
Label: Code-Review=+1