Update patch set 1
Patch Set 1:

In fact it looks like we already use github unless we use -g, which we can't do anymore apparently. This patch would only make a difference if you *did* supply a -g. It doesn't fix auth'ing to gerrit to be able to see a draft that an anonymous user wouldn't be able to see.

Patch-set: 1