Update patch set 2
Patch Set 2:

Could somebody with more experience on the review board please explain to me why being "against the patch" is even a thing?
'Tis but a simple bugfix. It does not introduce any new feature, so I don't get the debate on whether the feature is desirable or not, and it also has no side-effects if you're not using the feature.

Patch-set: 2
Reviewer: Vladimir Oltean <18924@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>