Update patch set 2
Patch Set 2:

Nobody said to download patches off the review board and hoard them in a BOARD_PATCHES_DIRS in the device repo. Everybody gets that that's what "repopick" is for. But keep in mind that not every bringup project starts on the review board. As suggested by Rashed, this change is now clearly intended solely for UNOFFICIAL and EXPERIMENTAL builds. If anything, I find the "DO NOT MERGE" workflow a bit backwards, but maybe that's just me.
To respond to Ethan's use case question, take this display rotation patch: https://github.com/vladimiroltean/android_device_lenovo_YTX703-common/blob/cm-14.1/patches/0001-framework-fix-display-orientation.patch
Of course this patch isn't going to fly on the review board, but:
* ro.sf.hwrotation isn't straightforward to do, since it breaks sensors, touchscreen and camera orientation
* It spans 2 totally unrelated repos, that only happen to be related through the issue at hand (the display's natural rotation). I ask - is it tidier to keep forks of frameworks/base and packages/apps/CMParts than to have a single patch file in the device repo, along with the infrastructure to support having that?
I see the support for board-provided patches as a stop-gap measure during board bring-up (so it won't slow you down). There is a time for bring-up, and there is a time for making things tidy and pushing them on the review board. The last thing I'd want to do is make people feel bad for fixing stuff - which is what the reluctance of supporting board-provided patches even for UNOFFICIAL builds does.

Patch-set: 2
Reviewer: Vladimir Oltean <18924@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>