[HACK] Patch external/svox
The svox repo can't be forked due to the following:
https://github.com/github/dmca/blob/master/2014/2014-12-22-Cambridge-Mobile.md

On Android 7 we use cee78199bbfae81f54a40671db47096f5f32cda as
revision ( https://android.googlesource.com/platform/external/svox/+/cee78199bbfae81f54a40671db47096f5f32cda )
which does not work on Android 6 and below.

Create the bug-id (bug_69177126) as empty file and check if it exist
to decide if external/svox needs to be patched or not.
For later use (Android 4?) and use in vendor/extra 
also check for the PLATFORM_VERSION.

Change-Id: Id312672ec1a93d4e52b0f81ec8134419ca70ead4