Update patch set 4
Patch Set 4:

This patch is another approach at this, without hardcoding every patch. It uses the Bug-ID mentioned in the Commit Message of the patch to determine if the patch needs to be applied, which removed the need for the bugid-files. However, it isn't able to handle multiple Bug-IDs as of now.

The version check was removed because this commit wouldn't be ported to other Android versions without testing (and adapting, if needed) anyways.

Patch-set: 4
Reviewer: Tim Schumacher <15734@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>