Update patch set 4
Patch Set 4:

This is just a friendly reminder to anyone else who comes along and sees this - this is an open source project, not a place to try to convince people your political views are correct. I've already removed a few comments/users, I would like to not have to remove any more. Thanks!

Patch-set: 4
CC: Tom Powell <5008@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>