Update patch set 1
Patch Set 1:

(1 comment)

The dictionaries are the same as added to 16.0 branch. On the same patch for 15.1 the cherry-pick worked but only on this patch not. The previous patch for polish dict files worked also. Should this be committed one again with a proper cherry-pick or can this patch be changed now?

Patch-set: 1
Reviewer: Gerrit User 22730 <22730@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>